Refactored app.js and split the functionality into smaller components based on their functionality
