+++ src/App.css
@@ -22,6 +22,11 @@ body {
   width: 450px;
 }
 
+#pokedex {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
 /* Search form */
 
 .searchField {
