+++ src/components/Header.jsx
@@ -1,0 +1,9 @@
+function Header() {
+    return (
+        <header>
+            <h1 className="title">Pokedex</h1>
+        </header>
+    );
+}
+
+export default Header;
